Implement scheduler changes for volume limits
